openrouter-multi-model/pyproject.toml
[project]
name = "openrouter-multi-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "httpx>=0.28.1",
    "openrouter-client>=0.1.0",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["uv"]
build-backend = "uv.build"
